feat: keep same module structure for generated python code
